[fixed] core3 client login connection
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@1700 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/creature/LuaCreatureObject.cpp
@@ -33,8 +33,8 @@ int LuaCreatureObject::_setObject(lua_State* L) {
 }
 
 int LuaCreatureObject::setHAM(lua_State* L) {
-	uint32 type = lua_tonumber(L, -2);
-	uint32 value = lua_tonumber(L, -1);
+	uint32 type = (uint32) lua_tonumber(L, -2);
+	uint32 value = (uint32) lua_tonumber(L, -1);
 
 	realObject->setHAM(type, value);
 
@@ -42,7 +42,7 @@ int LuaCreatureObject::setHAM(lua_State* L) {
 }
 
 int LuaCreatureObject::setBankCredits(lua_State *L) {
-	uint32 credits = lua_tonumber(L, -1);
+	uint32 credits = (uint32) lua_tonumber(L, -1);
 
 	realObject->setBankCredits(credits);
